Legg til minbredde på kolonner og splitt opp knapper
