+++ src/components/function-card-selected-view.tsx
@@ -3,7 +3,8 @@ import { MetadataView } from "./metadata/metadata-view";
 import { useMetadata } from "@/hooks/use-metadata";
 import { Route } from "@/routes";
 import { Button, Flex, Skeleton, Stack, Text } from "@kvib/react";
-import { EditAndSelectButtons } from "./edit-and-select-buttons";
+import { SelectButton } from "./buttons/select-button.tsx";
+import { EditButton } from "@/components/buttons/edit-button.tsx";
 
 export function FunctionCardSelectedView({
 	functionId,
@@ -16,7 +17,7 @@ export function FunctionCardSelectedView({
 	return (
 		<Stack pl="10px" w="100%" overflow="hidden">
 			<Flex alignItems="center" w="100%" flex-wrap="wrap">
-				<Skeleton isLoaded={!func.isLoading} flex={1} minWidth={0}>
+				<Skeleton isLoaded={!func.isLoading} flex={1} minWidth={10}>
 					<Text
 						fontWeight="bold"
 						as="span"
@@ -27,7 +28,10 @@ export function FunctionCardSelectedView({
 						{func.data?.name ?? "<Det skjedde en feil>"}
 					</Text>
 				</Skeleton>
-				<EditAndSelectButtons functionId={functionId} selected />
+				<Flex gap={1}>
+					<EditButton functionId={functionId} />
+					<SelectButton functionId={functionId} selected={false} />
+				</Flex>
 			</Flex>
 			{config.metadata?.map((meta) => {
 				const hasMetadata = metadata.data?.some((m) => m.key === meta.key);
